* rename snake.h to main.c * rename snake.h to game.h
